Remove delete button from general form error
Since bulma doesn't ship with javascript included, the delete/close button does not function. On top of this the delete/close will actually submit the form when clicked. This means the button does not only fail to perform its expected behavior, it also has unintended side effects. It's probably best that the button is removed or that a general message header replaces the button. 

This pull request just removes the malfunctioning button.